GeoAgent/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
argcomplete==3.5.1
asttokens==3.0.0
attrs==24.2.0
black==24.10.0
blinker==1.9.0
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datamodel-code-generator==0.26.3
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
folium==0.18.0
fonttools==4.55.2
frozenlist==1.5.0
future==1.0.0
genson==1.3.0
geojson==2.5.0
geojsplit==0.1.2
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
gpudb==7.2.2.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
ijson==2.6.1
inflect==5.6.2
ipykernel==6.29.5
ipython==8.30.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-community==0.2.17
langchain-core==0.2.43
langchain-experimental==0.0.62
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
mapclassify==2.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.56.0
orjson==3.10.12
osmnx==2.0.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.40.2
streamlit_folium==0.23.2
streamlit-iframe-event==0.0.9
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
xyzservices==2024.9.0
yarl==1.18.3
